requirements.txt
# ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# Deep Learning for Saliency Prediction
tensorflow>=2.13.0
keras>=2.13.0

# Image Processing and Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-image>=0.21.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0
h5py>=3.9.0

# Academic Dataset Tools
scikit-learn>=1.3.0
albumentations>=1.3.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 